Copied files from working directory
Allows users to be able to decide what type of graph they would like to display on the home page. As well modify the appetence of the home page. This is the extent of what we can do without access to their database with student records to display on the graphs